SCM sync has an option to display its status in the Jenkins  footer, and if an error occurs, a hyperlink is displayed to allow an admin  user to clear the error by calling the javascript:removeLog() function in the  footer.jelly file.
I noticed that the text containing the hyperlink was being displayed literally in the Jenkins footer, meaning that the link wasn't clickable and the error couldn't be cleared. It turns out that hyperlinks (and line breaks) are not being interpreted when contained in jelly string variables.

These modifications fix this issue.

From @s-p-turner
Fixes #5